+++ branches/devel-64bit/src/colinux/arch/i386/utils.c
@@ -12,9 +12,9 @@
 
 #include "utils.h"
 
-unsigned long co_get_cr4(void)
+uintptr_t co_get_cr4(void)
 {
-	unsigned long cr4 = 0;
+	uintptr_t cr4 = 0;
 	asm("mov %%cr4, %0" : "=r"(cr4));
 	return cr4;
 }
@@ -24,51 +24,51 @@ bool_t co_is_pae_enabled(void)
 	return co_get_cr4() & 0x20;
 }
 
-unsigned long co_get_cr3(void)
+uintptr_t co_get_cr3(void)
 {
-	unsigned long cr3 = 0;
+	uintptr_t cr3 = 0;
 	asm("mov %%cr3, %0" : "=r"(cr3));
 	return cr3;
 }
 
-unsigned long co_get_dr0(void)
+uintptr_t co_get_dr0(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr0, %0" : "=r"(reg));
 	return reg;
 }
 
-unsigned long co_get_dr1(void)
+uintptr_t co_get_dr1(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr1, %0" : "=r"(reg));
 	return reg;
 }
 
-unsigned long co_get_dr2(void)
+uintptr_t co_get_dr2(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr2, %0" : "=r"(reg));
 	return reg;
 }
 
-unsigned long co_get_dr3(void)
+uintptr_t co_get_dr3(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr3, %0" : "=r"(reg));
 	return reg;
 }
 
-unsigned long co_get_dr6(void)
+uintptr_t co_get_dr6(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr6, %0" : "=r"(reg));
 	return reg;
 }
 
-unsigned long co_get_dr7(void)
+uintptr_t co_get_dr7(void)
 {
-	unsigned long reg = 0;
+	uintptr_t reg = 0;
 	asm("mov %%dr7, %0" : "=r"(reg));
 	return reg;
 }
