+++ branches/devel-64bit/src/colinux/kernel/transfer.c
@@ -21,12 +21,12 @@
 co_rc_t co_monitor_host_linuxvm_transfer_map(
 	co_monitor_t*	cmon,
 	vm_ptr_t	vaddr,
-	unsigned long 	size,
+	uintptr_t 	size,
 	unsigned char** start,
 	unsigned char** page,
 	co_pfn_t*	ppfn
 ) {
-	unsigned long one_copy;
+	uintptr_t one_copy;
 	co_rc_t rc;
 
 	if ((vaddr < CO_ARCH_KERNEL_OFFSET) || (vaddr >= cmon->end_physical)) {
@@ -41,7 +41,7 @@ co_rc_t co_monitor_host_linuxvm_transfer_map(
 
 	one_copy = ((vaddr + CO_ARCH_PAGE_SIZE) & CO_ARCH_PAGE_MASK) - vaddr;
 	if (size <= 0 || size > one_copy) {
-		co_debug_error("monitor: transfer map: bad size: %ld (%ld)", size, one_copy);
+		co_debug_error("monitor: transfer map: bad size: %I64d (%I64d)", (int64_t)size, (int64_t)one_copy);
 		return CO_RC(TRANSFER_OFF_BOUNDS);
 	}
 
@@ -68,13 +68,13 @@ co_rc_t co_monitor_host_linuxvm_transfer(
 	void *host_data,
 	co_monitor_transfer_func_t host_func,
 	vm_ptr_t vaddr,
-	unsigned long size,
+	uintptr_t size,
 	co_monitor_transfer_dir_t dir
 	)
 {
 	co_pfn_t pfn;
 	unsigned char *page;
-	unsigned long one_copy;
+	uintptr_t one_copy;
 	co_rc_t rc;
 
 	if ((vaddr < CO_ARCH_KERNEL_OFFSET) || (vaddr >= cmon->end_physical)) {
@@ -112,7 +112,7 @@ co_rc_t co_monitor_host_linuxvm_transfer(
 
 
 static co_rc_t co_monitor_transfer_memcpy(co_monitor_t *cmon, void *host_data, void *linuxvm,
-					  unsigned long size, co_monitor_transfer_dir_t dir)
+					  uintptr_t size, co_monitor_transfer_dir_t dir)
 {
 	unsigned char **host = (unsigned char **)host_data;
 
@@ -127,7 +127,7 @@ static co_rc_t co_monitor_transfer_memcpy(co_monitor_t *cmon, void *host_data, v
 }
 
 static co_rc_t co_monitor_host_linuxvm_copy(co_monitor_t *cmon, void *host, vm_ptr_t linuxvm,
-					    unsigned long size, co_monitor_transfer_dir_t dir)
+					    uintptr_t size, co_monitor_transfer_dir_t dir)
 {
 	return co_monitor_host_linuxvm_transfer(cmon, &host, co_monitor_transfer_memcpy,
 						linuxvm, size, dir);
@@ -135,14 +135,14 @@ static co_rc_t co_monitor_host_linuxvm_copy(co_monitor_t *cmon, void *host, vm_p
 
 
 co_rc_t co_monitor_host_to_linuxvm(co_monitor_t *cmon, void *from,
-				   vm_ptr_t to, unsigned long size)
+				   vm_ptr_t to, uintptr_t size)
 {
 	return co_monitor_host_linuxvm_copy(cmon, from, to, size, CO_MONITOR_TRANSFER_FROM_HOST);
 }
 
 
 co_rc_t co_monitor_linuxvm_to_host(co_monitor_t *cmon, vm_ptr_t from,
-				   void *to, unsigned long size)
+				   void *to, uintptr_t size)
 {
 	return co_monitor_host_linuxvm_copy(cmon, to, from, size, CO_MONITOR_TRANSFER_FROM_LINUX);
 }
