+++ branches/devel-64bit/src/colinux/os/kernel/filesystem.h
@@ -27,14 +27,14 @@ extern co_rc_t co_os_fs_dir_inode_to_path(co_filesystem_t *fs, co_inode_t *dir,
  * OS-specific operations on files.
  */
 extern co_rc_t co_os_file_read_write(struct co_monitor *linuxvm, char *filename, unsigned long long offset,
-				     unsigned long size, vm_ptr_t src_buffer, bool_t read);
-extern co_rc_t co_os_file_set_attr(char *filename, unsigned long valid, struct fuse_attr *attr);
+				     uintptr_t size, vm_ptr_t src_buffer, bool_t read);
+extern co_rc_t co_os_file_set_attr(char *filename, uintptr_t valid, struct fuse_attr *attr);
 extern co_rc_t co_os_file_get_attr(char *filename, struct fuse_attr *attr);
 extern co_rc_t co_os_file_unlink(char *filename);
 extern co_rc_t co_os_file_rmdir(char *filename);
 extern co_rc_t co_os_file_mkdir(char *dirname);
 extern co_rc_t co_os_file_rename(char *filename, char *dest_filename);
-extern co_rc_t co_os_file_mknod(co_filesystem_t *filesystem, char *filename, unsigned long mode);
+extern co_rc_t co_os_file_mknod(co_filesystem_t *filesystem, char *filename, uintptr_t mode);
 extern co_rc_t co_os_file_getdir(char *dirname, co_filesystem_dir_names_t *names);
 extern co_rc_t co_os_file_fs_stat(co_filesystem_t *filesystem, struct fuse_statfs_out *statfs);
